test: working on fixing the tests
